Test some implementation validation
