+++ lib/graphql/schema/implementation.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
+require "graphql/schema/implementation/apply_proxies"
+require "graphql/schema/implementation/build_resolve"
+require "graphql/schema/implementation/invalid_implementation_error"
 require "graphql/schema/implementation/method_call_implementation"
+require "graphql/schema/implementation/proxied_object"
 require "graphql/schema/implementation/public_send_implementation"
 require "graphql/schema/implementation/type_missing"
 
@@ -57,87 +61,14 @@ module GraphQL
 
       private
 
-      # TODO can this be applied in initialization?
-      class ApplyProxies
-        def initialize(proxy_map)
-          @proxy_map = proxy_map
-        end
-
-        def instrument(type, field)
-          owner_type_name = type.name
-          # TODO don't hardcode
-          resolve_proc = if owner_type_name == "Query" || owner_type_name == "Mutation"
-            old_resolve = field.resolve_proc
-            ->(o, a, c) {
-              proxy_cls = @proxy_map[owner_type_name]
-              proxy = proxy_cls.new(o, c)
-              proxied_object = ProxiedObject.new(o, proxy)
-              old_resolve.call(proxied_object, a, c)
-            }
-          else
-            field.resolve_proc
-          end
-
-          inner_return_type = field.type.unwrap
-          resolve_proc_w_proxy = case inner_return_type
-          when GraphQL::ObjectType, GraphQL::InterfaceType, GraphQL::UnionType
-            depth = list_depth(field.type)
-            ->(o, a, c) {
-              result = resolve_proc.call(o, a, c)
-              proxy_to_depth(result, depth, inner_return_type, @proxy_map, c)
-            }
-          else
-            resolve_proc
-          end
-
-          field.redefine(resolve: resolve_proc_w_proxy)
-        end
-
-        private
-
-        def list_depth(type, starting_at = 0)
-          case type
-          when GraphQL::ListType
-            list_depth(type.of_type, starting_at + 1)
-          when GraphQL::NonNullType
-            list_depth(type.of_type, starting_at)
-          else
-            starting_at
-          end
-        end
-
-        def proxy_to_depth(obj, depth, type, proxies, ctx)
-          if depth > 0
-            obj.map { |inner_obj| proxy_to_depth(inner_obj, depth - 1, type, proxies, ctx) }
-          else
-            concrete_type = type.resolve_type(obj, ctx)
-            proxy_cls = proxies[concrete_type.name]
-            proxy = proxy_cls.new(obj, ctx)
-            ProxiedObject.new(obj, proxy)
-          end
-        end
-      end
-
-      class ProxiedObject
-        attr_reader :object, :proxy
-        def initialize(object, proxy)
-          @object = object
-          @proxy = proxy
-        end
-      end
-
-      # TODO, can users extend this, provide custom args?
-      SPECIAL_ARGS = Set.new([
-        :context,
-        :irep_node,
-        :ast_node,
-      ])
-
       def build_fields(schema)
+        # TODO this is a stupid hack, sharing the same hash like this
         map = {
           proxies: {}
         }
 
+        # This is a fallback in case there's not a proxy class defined.
+        # Application can define its own fallback, or we provide one
         default_impl = if @namespace.const_defined?(:TypeMissing)
           @namespace.const_get(:TypeMissing)
         else
@@ -146,7 +77,11 @@ module GraphQL
 
         schema.types.each do |name, type|
           if type.kind.fields?
+            # Introspection types can be defined in Introspection namespace
+            # (constant names don't have `__`)
             impl_constant_name = name.sub(/^__/, "Introspection::")
+
+            # Try to find a defined proxy class, otherwise use the fallback
             impl_class = if @namespace.const_defined?(impl_constant_name)
               @namespace.const_get(impl_constant_name)
             else
@@ -156,43 +91,8 @@ module GraphQL
             fields = map[name] = {}
 
             type.all_fields.each do |field|
-              field_name = field.name
-              field_args = field.arguments
-              # Remove camelization
-              method_name = field_name
-                .gsub(/([A-Z]+)([A-Z][a-z])/,'\1_\2')
-                .gsub(/([a-z\d])([A-Z])/,'\1_\2')
-                .downcase
-
-              fields[field.name] = if impl_class.method_defined?(method_name)
-                field_method = impl_class.instance_method(method_name)
-                graphql_args = []
-                special_args = []
-                # TODO assert required-ness matches between Ruby & GraphQL
-                field_method.parameters.each do |(type, name)|
-                  case type
-                  when :req, :opt
-                    raise "Positional arguments are not supported, see #{type_name}.#{method_name} (#{field_method.source_location})"
-                  when :keyreq, :key
-                    arg_name = name.to_s
-                    if field_args.key?(arg_name)
-                      graphql_args << name
-                    elsif SPECIAL_ARGS.include?(name)
-                      special_args << name
-                    else
-                      raise "#{type_name}.#{method_name}: Unexpected argument name #{name.inspect}, expected one of #{(field_args.keys + SPECIAL_ARGS).map(&:to_sym).map(&:inspect)} (#{field_method.source_location})"
-                    end
-                  end
-                end
-
-                Implementation::MethodCallImplementation.new(
-                  method: method_name,
-                  graphql_arguments: graphql_args,
-                  special_arguments: special_args,
-                )
-              else
-                Implementation::PublicSendImplementation.new(method: method_name)
-              end
+              # Build a field from this proxy class
+              fields[field.name] = BuildResolve.build(impl_class, field)
             end
           end
         end
