+++ lib/graphql/schema/implementation/invalid_implementation_error.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+module GraphQL
+  class Schema
+    class Implementation
+      class InvalidImplementationError < GraphQL::Error
+      end
+    end
+  end
+end
